# Faster `glyf` compile