Mostly around `recalcBounds`.

For Handjet font, this bring compile time down from 11s to 0.22s:
```
time ./fonttools ttLib Handjet.ttf -o out.ttf -tglyf
```

Brings down instancing speed for me from 23s down to 0.36s.

Part of fixing https://github.com/fonttools/fontbakery/discussions/4301